Fix compression/decompression performance issue on windows
Apparently, we weren't closing the GZip stream.
This shouldn't be an issue as there really isn't any underlying file or other native resource backing it, just an Inflater/Deflater.
Their documentation says that end() (called by stream close) releases the underlying resources. And it says that it's automatically called from finalizer.
Except on windows it isn't because the object ends up being leaked and never GCed.